create file with news module protocols
